moved getTasksOfProperty functions into separate controller
